+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## [0.2.2]
+- relations are now base::types::Relation and are a Vec<Ulid> to support more complex relations, big
+breaking change for everything
+
 ## [0.2.1]
 - split load and load_dyn on query in preparation for inter-canister queries
-- relations are now base::types::Relation and are a Vec<Ulid> to support more complex relations
 
 ## [0.2.0]
 - all changes that make mimic compatible with the Dragginz repo v0.2